@@ -4,39 +4,43 @@
     "version": "2.0.0",
     "type": "module",
     "scripts": {
-        "check:types": "tsc --noEmit --skipLibCheck",
-        "start": "vite",
-        "dev": "vite --host",
         "build": "tsc && vite build",
+        "check": "pnpm dlx @biomejs/biome check",
+        "check:all": "pnpm dlx @biomejs/biome check .",
+        "check:fix:all": "pnpm dlx @biomejs/biome check --apply .",
+        "check:format": "pnpm dlx @biomejs/biome format",
+        "check:format:all": "pnpm dlx @biomejs/biome format .",
+        "check:format:fix": "pnpm dlx @biomejs/biome format --write",
+        "check:format:fix:all": "pnpm dlx @biomejs/biome format --write .",
+        "check:lint": "pnpm dlx @biomejs/biome lint",
+        "check:lint:fix": "pnpm dlx @biomejs/biome lint --apply",
+        "check:lint:fix:all": "pnpm dlx @biomejs/biome lint --apply .",
+        "check:lint:all": "pnpm dlx @biomejs/biome lint .",
+        "check:types": "tsc --noEmit --skipLibCheck",
+        "dev": "vite --host",
         "preview": "vite preview",
-        "lint": "eslint . --ext .ts,.tsx",
-        "lint:print-config": "eslint --print-config file.js > eslintconfig-out.json"
+        "start": "vite"
     },
     "dependencies": {
         "@fortawesome/free-brands-svg-icons": "^6.4.0",
         "@fortawesome/free-regular-svg-icons": "^6.4.0",
         "@fortawesome/free-solid-svg-icons": "^6.4.0",
-        "@motionone/solid": "^10.16.2",
         "@solidjs/meta": "^0.28.6",
         "@solidjs/router": "^0.8.2",
         "lodash.throttle": "^4.1.1",
         "solid-fa": "^0.2.0",
-        "solid-js": "^1.7.9"
+        "solid-js": "^1.8.10",
+        "solid-motionone": "^1.0.0"
     },
     "devDependencies": {
-        "@pxeeio/eslint-config-solid": "^0.1.2",
+        "@biomejs/biome": "1.5.0",
         "@types/lodash.throttle": "^4.1.7",
         "@types/node": "^18.15.10",
-        "@typescript-eslint/eslint-plugin": "^6.2.1",
-        "@typescript-eslint/parser": "^6.2.1",
         "autoprefixer": "^10.4.14",
-        "eslint": "^8.46.0",
-        "eslint-define-config": "^1.22.0",
         "postcss": "^8.4.27",
-        "prettier": "^2.8.7",
         "sass": "^1.64.2",
-        "tailwindcss": "^3.3.3",
-        "typescript": "^5.0.2",
+        "tailwindcss": "^3.4.1",
+        "typescript": "^5.3.3",
         "typescript-plugin-css-modules": "^5.0.1",
         "vite": "^4.4.8",
         "vite-plugin-solid": "^2.7.0"
