@@ -1,7 +1,7 @@
-import {Presence} from '@motionone/solid';
 import {MetaProvider} from '@solidjs/meta';
 import {Router, useRoutes} from '@solidjs/router';
-import {type Component, onMount} from 'solid-js';
+import type {Component} from 'solid-js';
+import {Presence} from 'solid-motionone';
 
 import routes from '@/navigation/routes';
 import {initializeLocalStore} from '@/stores/local.store';
@@ -9,9 +9,7 @@ import {initializeLocalStore} from '@/stores/local.store';
 const Root: Component = () => {
     const Routes = useRoutes(routes);
 
-    onMount(() => {
-        initializeLocalStore();
-    });
+    initializeLocalStore();
 
     return (
         <MetaProvider>
