@@ -1,16 +1,40 @@
 import {Link, Meta, Title} from '@solidjs/meta';
-import {type Component, Index, mergeProps, Show, splitProps} from 'solid-js';
+import {
+    type Component,
+    type ComponentProps,
+    Index,
+    type JSX,
+    Show,
+    mergeProps,
+    splitProps,
+} from 'solid-js';
 
 import OpenGraphHelmet from '@/components/Helmet/OpenGraphHelmet';
 import TwitterCardHelmet from '@/components/Helmet/TwitterCardHelmet';
-import type {
-    LinkAttribute,
-    MetaAttribute,
-    OpenGraphAttributes,
-    TwitterCardAttributes,
-} from '@/components/Helmet/types';
 
-export type HelmetProps = {
+type LinkAttribute =
+    | {
+          rel: 'preload';
+          as: JSX.IntrinsicElements['link']['as'];
+          type: `text/${'css' | 'html' | 'js'}`;
+      }
+    | {
+          rel: string;
+          type: string;
+          [key: string]: string;
+      };
+
+type MetaAttribute =
+    | {
+          name: string;
+          content: string;
+      }
+    | {
+          property: string;
+          content: string;
+      };
+
+type HelmetProps = {
     author?: string;
     description?: string;
     favicon?: `${string}.ico`;
@@ -18,9 +42,10 @@ export type HelmetProps = {
     title?: string;
     links?: LinkAttribute[];
     meta?: MetaAttribute[];
-} & OpenGraphAttributes & TwitterCardAttributes;
+} & ComponentProps<typeof OpenGraphHelmet> &
+    ComponentProps<typeof TwitterCardHelmet>;
 
-const Helmet: Component<HelmetProps> = props => {
+const Helmet: Component<HelmetProps> = (props) => {
     const propsWithDefaults = mergeProps(
         {
             links: [],
@@ -50,31 +75,31 @@ const Helmet: Component<HelmetProps> = props => {
     return (
         <>
             <Show when={remainingProps.title} keyed>
-                {title => <Title>{title}</Title>}
+                {(title) => <Title>{title}</Title>}
             </Show>
 
             <Show when={remainingProps.favicon} keyed>
-                {favicon => <Link rel="icon" type="image/ico" href={favicon} />}
+                {(favicon) => <Link rel="icon" type="image/ico" href={favicon} />}
             </Show>
 
             <Show when={remainingProps.author} keyed>
-                {author => <Meta name="author" content={author} />}
+                {(author) => <Meta name="author" content={author} />}
             </Show>
 
             <Show when={remainingProps.description} keyed>
-                {description => <Meta name="description" content={description} />}
+                {(description) => <Meta name="description" content={description} />}
             </Show>
 
             <Show when={remainingProps.image} keyed>
-                {image => <Meta name="image" content={image} />}
+                {(image) => <Meta name="image" content={image} />}
             </Show>
 
             <OpenGraphHelmet {...openGraphProps} />
 
             <TwitterCardHelmet {...twitterCardProps} />
-            
-            <Index each={remainingProps.links}>{link => <Link {...link()} />}</Index>
-            <Index each={remainingProps.meta}>{meta => <Meta {...meta()} />}</Index>
+
+            <Index each={remainingProps.links}>{(link) => <Link {...link()} />}</Index>
+            <Index each={remainingProps.meta}>{(meta) => <Meta {...meta()} />}</Index>
         </>
     );
 };
